refactor: remove unused components and streamline post page structure
