--- src/app/post/[pubky]/[postId]/components/index.tsx
@@ -1,19 +1,0 @@
-import { NavigatorParent } from './_NavigatorParent';
-import { Participants } from './_Participants';
-import { PostRoot } from './_PostRoot';
-import { Replies } from './_Replies';
-import { RootParent } from './_RootParent';
-import { ValidPostContent } from './_ValidPostContent';
-import { Root } from './_Root';
-
-export const Post = {
-  NavigatorParent,
-  Participants,
-  PostRoot,
-  Replies,
-  RootParent,
-  ValidPostContent,
-  Root
-};
-
-export default Post;
